requirements.txt
ariadne==0.13.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
graphql-core==3.1.6
greenlet==1.1.2
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
loguru==0.5.3
MarkupSafe==2.0.1
packaging==21.0
pinecone-client==2.0.1
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
sentry-sdk==1.4.3
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
